+++ src/ledger/settings.ts
@@ -4,6 +4,7 @@ import {validate, constants} from '../common'
 import {FormattedSettings} from '../common/types/objects'
 import {AccountInfoResponse} from '../common/types/commands'
 import {RippleAPI} from '..'
+import { xAddressToClassicAddress } from 'ripple-address-codec'
 const AccountFlags = constants.AccountFlags
 
 export type SettingsOptions = {
@@ -38,6 +39,22 @@ export async function getSettings(
 ): Promise<FormattedSettings> {
   // 1. Validate
   validate.getSettings({address, options})
+
+  // Only support retrieving settings without a tag,
+  // since settings do not distinguish by tag.
+  try {
+    const {
+      classicAddress,
+      tag
+    } = xAddressToClassicAddress(address)
+    if (tag !== false) {
+      return Promise.reject('getSettings does not support the use of a tag. Use an address without a tag.')
+    }
+    address = classicAddress
+  } catch (_) {
+    // `address` is already a classic address
+  }
+
   // 2. Make Request
   const response = await this.request('account_info', {
     account: address,