+++ src/ledger/trustlines.ts
@@ -3,6 +3,7 @@ import {validate} from '../common'
 import parseAccountTrustline from './parse/account-trustline'
 import {RippleAPI} from '..'
 import {FormattedTrustline} from '../common/types/objects/trustlines'
+import { xAddressToClassicAddress } from 'ripple-address-codec'
 
 export type GetTrustlinesOptions = {
   counterparty?: string,
@@ -20,11 +21,27 @@ async function getTrustlines(
 ): Promise<FormattedTrustline[]> {
   // 1. Validate
   validate.getTrustlines({address, options})
-  const ledgerVersion = await this.getLedgerVersion()
+
+  // Only support retrieving trustlines without a tag,
+  // since it does not make sense to filter trustlines
+  // by tag.
+  try {
+    const {
+      classicAddress,
+      tag
+    } = xAddressToClassicAddress(address)
+    if (tag !== false) {
+      return Promise.reject('getTrustlines does not support the use of a tag. Use an address without a tag.')
+    }
+    address = classicAddress
+  } catch (_) {
+    // `address` is already a classic address
+  }
+
   // 2. Make Request
   const responses = await this._requestAll('account_lines', {
     account: address,
-    ledger_index: ledgerVersion,
+    ledger_index: await this.getLedgerVersion(),
     limit: options.limit,
     peer: options.counterparty
   })